pyproject.toml
[tool.poetry]
name = "resellers-wrapper"
version = "0.1.0"
description = ""
authors = ["Marcellort <marcellosantana@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tapioca-wrapper = "^2.3.0"
pydantic-settings = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
